+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased][beta]
+## [Unreleased][latest]
 ### Added
 - AWS Lambda Timeout Warning (#679)
 - Enforce Conventional Commits using a git hook (#731)