+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][latest]
+
+## [4.0.0] - 2022-05-04
 ### Added
 - AWS Lambda Timeout Warning (#679)
 - Enforce Conventional Commits using a git hook (#731)